requirements.txt
langchain==0.2.11
langchain-community==0.2.10
langchain-huggingface==0.0.3
langchain-chroma==0.1.2
chromadb==0.5.5
sentence_transformers==3.0.1
requests==2.32.3
python-dotenv==1.0.1
langchain-groq==0.1.8
grpcio-tools==1.65.1